refactor(federated-logs): segregate unit tests for aws connection
